API_server/requirements.txt
connexion >= 2, < 3
connexion[swagger-ui] >= 2, < 3
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2

Flask~=2.2.2
six~=1.16.0
kubernetes >= 10.0.0
